pyproject.toml
[project]
name = "divina-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "divina-classifier",
    "flask==3.0.3",
    "flask-bcrypt==1.0.1",
    "flask-sqlalchemy==3.1.1",
    "pillow>=11.0.0",
    "pydantic>=2.12.5",
    "pyjwt==2.9.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "sqlalchemy==2.0.36",
]

[tool.uv.sources]
divina-classifier = { git = "https://github.com/Team-Kalubihan/DIVINA-Classifier.git" }
